GraphQL_API/pyproject.toml
[project]
name = "graphql-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.14",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.41",
    "strawberry-graphql[fastapi]>=0.275.5",
    "strawberry-sqlalchemy-mapper>=0.6.3",
    "typer>=0.16.0",
]

[dependency-groups]
dev = [
    "debugpy>=1.8.14",
    "ruff>=0.12.2",
]
